+++ src/ACEtk/photonuclear/PrincipalCrossSectionBlock/test/CMakeLists.txt
@@ -1,6 +1,6 @@
 
-add_executable( ACEtk.block.PhotonuclearSecondaryParticleLocatorBlock.test PhotonuclearSecondaryParticleLocatorBlock.test.cpp )
-target_compile_options( ACEtk.block.PhotonuclearSecondaryParticleLocatorBlock.test PRIVATE ${${PREFIX}_common_flags}
+add_executable( ACEtk.photonuclear.PrincipalCrossSectionBlock.test PrincipalCrossSectionBlock.test.cpp )
+target_compile_options( ACEtk.photonuclear.PrincipalCrossSectionBlock.test PRIVATE ${${PREFIX}_common_flags}
 $<$<BOOL:${strict}>:${${PREFIX}_strict_flags}>$<$<CONFIG:DEBUG>:
 ${${PREFIX}_DEBUG_flags}
 $<$<BOOL:${coverage}>:${${PREFIX}_coverage_flags}>>
@@ -10,9 +10,9 @@ $<$<BOOL:${link_time_optimization}>:${${PREFIX}_link_time_optimization_flags}>
 $<$<BOOL:${nonportable_optimization}>:${${PREFIX}_nonportable_optimization_flags}>>
 
 ${CXX_appended_flags} ${ACEtk_appended_flags} )
-target_link_libraries( ACEtk.block.PhotonuclearSecondaryParticleLocatorBlock.test PUBLIC ACEtk )
+target_link_libraries( ACEtk.photonuclear.PrincipalCrossSectionBlock.test PUBLIC ACEtk )
 file( GLOB resources "resources/*" )
 foreach( resource ${resources})
     file( COPY "${resource}" DESTINATION "${CMAKE_CURRENT_BINARY_DIR}" )
 endforeach()
-add_test( NAME ACEtk.block.PhotonuclearSecondaryParticleLocatorBlock COMMAND ACEtk.block.PhotonuclearSecondaryParticleLocatorBlock.test )
+add_test( NAME ACEtk.photonuclear.PrincipalCrossSectionBlock COMMAND ACEtk.photonuclear.PrincipalCrossSectionBlock.test )
