+++ src/ACEtk/photonuclear/PrincipalCrossSectionBlock/test/PrincipalCrossSectionBlock.test.cpp
@@ -1,7 +1,7 @@
 #define CATCH_CONFIG_MAIN
 
 #include "catch.hpp"
-#include "ACEtk/block/PhotonuclearPrincipalCrossSectionBlock.hpp"
+#include "ACEtk/photonuclear/PrincipalCrossSectionBlock.hpp"
 
 // other includes
